tests/robot/requirements.txt
robotframework>=7.0,<8.0
robotframework-requests>=0.9.7,<1.0
